+++ NEWS.md
@@ -1,3 +1,10 @@
+# Version 0.2.0
+
+* New filters tool from <https://editor.nomics.world/filters>.
+* If the retrieved dataset contains columns with codes (like ISO codes,
+  geographic codes, ...), then correspondences are performed to translate
+  these codes if possible.
+
 # Version 0.1.0
 
 * First release.